removed screen dependency from router, it uses a simple callback now
