+++ frontend/app/context.go
@@ -4,6 +4,7 @@ import (
 	"godl/backend/events"
 	"godl/backend/model"
 	"godl/frontend/components"
+	"godl/frontend/router"
 
 	"github.com/rivo/tview"
 )
@@ -11,6 +12,7 @@ import (
 type AppContext struct {
 	app    *tview.Application
 	navBar *components.NavBarComponent
+	router *router.Router
 }
 
 func (ac *AppContext) initApp() {
@@ -21,9 +23,11 @@ func (ac *AppContext) addRoutes() {
 
 func NewAppContext() *AppContext {
 	a := tview.NewApplication()
+	r := router.NewRouter()
 	ac := &AppContext{
 		app:    a,
-		navBar: components.NewNavBarComponent(),
+		navBar: components.NewNavBarComponent(r),
+		router: r,
 	}
 	ac.addRoutes()
 	ac.setupNavBar()
