--- frontend/components/navBar/router.go
@@ -1,27 +1,0 @@
-package navBar
-
-import "godl/frontend/screens"
-
-type route string
-
-type Router struct {
-	routes map[route]screens.Screen
-}
-
-func NewRouter() *Router {
-	return &Router{
-		routes: make(map[route]screens.Screen),
-	}
-}
-
-func (m *Router) Draw(r route) {
-	m.routes[r].Draw()
-}
-
-func (m *Router) AddRoute(r route, s screens.Screen) {
-	m.routes[r] = s
-}
-
-func (m *Router) GetResolution(r route) func() {
-	return m.routes[r].Draw
-}
